Fixes and cleaning.
Fixed UI bug with loading icon appearing on hover.
Cleaned up whole project deleting all unnecessary code. There was a lot!
