--- README.md
@@ -1,7 +1,0 @@
-# Alfa
-Building
-
-
-## npm install
-
-## npm start
