+++ src/components/Airdrop.css
@@ -30,7 +30,7 @@
   margin-right: 16px;
 }
 .component1-button-loader:after {
-  content: ' ';
+  content: " ";
   display: block;
   width: 16px;
   height: 16px;
@@ -39,6 +39,10 @@
   border-color: var(--dl-color-template-themepurple) transparent
     var(--dl-color-template-themepurple) transparent;
   animation: component1-button-loader 1.2s linear infinite;
+}
+.component1-button:hover .component1-button-loader:after {
+  border: 2px solid white;
+  border-color: white transparent white transparent;
 }
 @keyframes component1-button-loader {
   0% {
