+++ src/components/Zkdrop/Zkdrop.tsx
@@ -1,15 +1,13 @@
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
+import React, { useState } from "react";
+import PropTypes from "prop-types";
 
-import {
-    bridgeSwap,
-} from './drop';
+import { bridgeSwap } from "./drop";
 
-import './Zkdrop.css';
+import "../Airdrop.css";
 
 const Zkdrop = (props: any) => {
   const [running, setRunning] = useState(false);
-  const [status, setStatus] = useState('Collect');
+  const [status, setStatus] = useState("Collect");
 
   return (
     <div className={`component1-container ${props.rootClassName} `}>
@@ -17,20 +15,18 @@ const Zkdrop = (props: any) => {
         onClick={() => {
           if (running === false) {
             setRunning(true);
-            (
-                async () => {
-                    try {
-                        await bridgeSwap(setStatus);
-                    } catch (e) {
-                        console.log(e);
-                        setStatus('Error encountered');
-                    }
+            (async () => {
+              try {
+                await bridgeSwap(setStatus);
+              } catch (e) {
+                console.log(e);
+                setStatus("Error encountered");
+              }
 
-                    setRunning(false);
-                }
-            )();
+              setRunning(false);
+            })();
           } else {
-            alert('Cannot run multiple times');
+            alert("Cannot run multiple times");
           }
         }}
         className="component1-button themebutton button"
@@ -45,7 +41,7 @@ const Zkdrop = (props: any) => {
 };
 
 Zkdrop.defaultProps = {
-  rootClassName: '',
+  rootClassName: "",
 };
 
 Zkdrop.propTypes = {
