--- src/components/component1.css
@@ -1,19 +1,0 @@
-.component1-container {
-  display: flex;
-  position: relative;
-}
-.component1-button {
-  flex: 1;
-  transition: 0.3s;
-}
-.component1-button:hover {
-  color: #ffffff;
-  border-color: #e2e2e2;
-  justify-content: center;
-  background-color: var(--dl-color-template-themepurple);
-}
-.component1-root-class-name {
-  left: 25px;
-  bottom: 33px;
-  position: static;
-}
