--- src/components/schedule.css
@@ -1,24 +1,0 @@
-.schedule-schedule {
-  position: relative;
-}
-.schedule-icon {
-  width: 32px;
-  height: 30px;
-  margin-bottom: 24px;
-}
-.schedule-text {
-  font-size: 28px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 33px;
-  margin-bottom: 16px;
-}
-.schedule-text1 {
-  line-height: 24px;
-}
-@media (max-width: 767px) {
-  .schedule-schedule {
-    margin-right: 0px;
-    margin-bottom: var(--dl-space-space-twounits);
-  }
-}
