--- src/components/you-tube.css
@@ -1,22 +1,0 @@
-.you-tube-you-tube {
-  width: 240px;
-  height: 120px;
-  display: flex;
-  align-items: center;
-  border-radius: 200px;
-  justify-content: center;
-  background-color: var(--dl-color-gray-white);
-}
-.you-tube-pasted-image {
-  width: 120px;
-}
-@media (max-width: 479px) {
-  .you-tube-you-tube {
-    width: 120px;
-    height: 60px;
-    align-items: center;
-  }
-  .you-tube-pasted-image {
-    width: 80px;
-  }
-}
