+++ src/views/DappGate.js
@@ -1,0 +1,529 @@
+import React, { useState } from "react";
+import { useContractRead, useAccount } from "wagmi";
+import contractABI from "../abis/abi.json";
+import { toast } from "react-toastify";
+// import { ethers } from "ethers";
+import { useNavigate } from "react-router-dom";
+import { Helmet } from "react-helmet";
+import { ConnectionButton } from "../components/ConnectButton";
+import "./DappGate.css";
+
+const DappGate = (props) => {
+  const { address, isConnected } = useAccount();
+
+  let navigate = useNavigate();
+
+  // Alfa Token mainnet address is the contract address here
+  const contractAddress = "0x128ad1ad707c3B36e6F2ac9739f9dF7516FdB592";
+  //ALphawolves NFT ca
+  const contractAddress2 = "0xdcd6d4a557ff208f01D4c2b5Bf829078622C37c5";
+
+  const { data: balanceOfCheck } = useContractRead({
+    address: contractAddress,
+    abi: contractABI,
+    functionName: "balanceOf",
+    args: [address],
+  });
+
+  const { data: balanceOfCheck2 } = useContractRead({
+    address: contractAddress2,
+    abi: contractABI,
+    functionName: "balanceOf",
+    args: [address],
+  });
+
+  const wl = [
+    "0xcB45248Aa2960aaDDA73a780efbC1fFd6B0D3338",
+    "0xCcFf390811bCdbf646f8dbABcD2efC467a846999",
+    "0x393d7CAfE78c6f80D57d741f31cdeAa77787754c",
+    "0xd89B6e2B3fdC0c73788b12C0cE4ba6425a7359C7",
+    "0xD70af4e0ECD011a1dad4C7f7a72A5f1FdB5D9C91",
+    "0xD771746b21345073E7D0F4c6b80A02F4cA131D9d",
+    "0x08774692eFF2f610f820E4E7C9d331411A3BBE38",
+    "0x5B3ed0486f87D21F70fEE88C58Bc31174247e2a9",
+    "0xE917c99290466B64bBCc9217E358025546b55Dd1",
+    "0x00A1ACFE4cE36F7e09919fe80d8CC2b3e2727896",
+    "0x27f360F7B12496004B148e76CDfeD3e0267d4761",
+    "0x93D35166cb3052eD0D4aE04cAA4DF4013E061018",
+    "0x870bFF1992020964046560f3EAECBf8f20714052",
+    "0x5Efc2bfc26235420A90cC330068C6bAd3dB394bd",
+    "0x704e7B754Ff03645379b8331B3F3D960913A25Dc",
+    "0x0486067B67ACe7CcBCE5FC93c7186923aE6b248A",
+    "0xC719f29A7483a098A07C01700144578BeAF08Ea5",
+    "0x6Da6d88Fe24885c27fAE65f58ad7868718Bf40DC",
+    "0xeA891A3699C81658B8291f60024d8B0bb435196f",
+    "0x881Cefe0791ce275Cc2415103d9491035e3904CA",
+    "0xc6daB5C113EB8647837caf4Bbfe9aB2cA8449544",
+    "0xEB00C0F8bB1DDf074B464DbCbE32164aD00E1a4B",
+    "0xEB40484c6a1BB5AD6531c3D59c06eD07Ae36030b",
+    "0x106d4BC771de874907D3d31F2129837963556318",
+    "0x573a5C5172A555F381Df3917ea9A30eaFC72998a",
+    "0x6654760eb57eBb2a841A363b61cFc36f6ea48B5b",
+    "0x7f2D15040336b0C84A14d6551ED582eB44b626FC",
+    "0x2787F81Ba8366DB54595767Ed4D1997B5bEDe140",
+    "0x3b5876896Dfa2089E67291E8CB4B7DF1c950619E",
+    "0xbFD082bA9aD6CF2976a794c4FB677a5349b301BA",
+    "0x517D72AE57E4CcAcAa7a9F3BFfF3C97613a5C6B0",
+    "0x941A8093C3d975326a3089f4df8228218Bf63698",
+    "0x0dCE9F4982f5A2277fB551AA538379fEaD465750",
+    "0x64EdA023719a6d7F706879C8fcDd71b860eBDb42",
+    "0xbBb4A122c710CD8906EBaC2dBF74Ed4C102C777E",
+    "0x2Af0FE506829eFf107a459403Dd4DD541FD7CE4E",
+    "0xBf87E7254D4fD063A69BF67De73790c43d468649",
+    "0xc6aA455a178B8dB2c9BFAFdDF726Cf1222d20a24",
+    "0x72D7a77a35eB75490Db7B0648051755bFe5CE42F",
+    "0x904CB35fC63FB65EfE38a7772309731dD0223353",
+    "0x6462154bd6Db79C6b6fF962d53a7810E3ebA7870",
+    "0x27973Bb46aE24c9802c9BC47ea938a8b76f343Db",
+    "0x4054C437c15cC8c194b253D09B725BE2b783D77b",
+    "0x4282d5C7F12624b66eb3e05CAB3b2530dF222553",
+    "0x189Fed2037Fa15d6feA1060BB3e634102239548d",
+    "0x4D2f5F90c69Ae60A794495A6a2B8D771B07572E2",
+    "0x818fDF253932f46fB6761c2A137ccbE2553C19e9",
+    "0x95C03aa406648961b63e371651ACd1ebeeD5186D",
+    "0xbEfaCcA046A02262E8D38314573d2bd466803e55",
+    "0x7611ba75C111bAd028d0cBD61cF4FaB6089E2860",
+    "0x9713e12F2265eFa8b1bCC99A07e287038f31BBC0",
+    "0x859F4cDeae8af3f714fb1a3993D8FcF600DA67b5",
+    "0xAAcE123ef3222F24234e46823Ca42B794F8eF392",
+    "0x3d4f4E4F3e012E1258448B1a07985e4288a8bf36",
+    "0xe4958f94014f87E1370404746Fd7944a7E72efc8",
+    "0x0FcF993690A7C27Dcbd1bdc7873dEbA7b54b2ac1",
+    "0x1a5940455D531672Af78982e164C4563dDBF40cB",
+    "0x936510f2bbF1eB485f5bE1eA60124AaD0D41D131",
+    "0x98DBC070AD90ce66094b0F108511C48D32E1A47f",
+    "0x4fC9529194ed01c4BfDaB22A5fE699d2e0E4b243",
+    "0x7754AB999455B8045901fda8f5440d5CF5743fEa",
+    "0x09BAA9a11fE737D5C5ADF149D590d7CE7F2b08b3",
+    "0x38BCa3Ac5f563454c38A93Ec45bB99232E63D500",
+    "0x6e40192503D849Ec8FB649CF8b0548FD28B3AB26",
+    "0xEfb54BE611B0B7157242418a69dc969772A544E5",
+    "0x0C6aED9d9C7e6de4c50fab6d722e490ea49b13Ee",
+    "0xCB00b2D04cEc567f03b37Ce6DA0fFBf5683f54De",
+    "0xf183DD1daBf4F33EbfD7392E4895e7e9A3383d3a",
+    "0xE8F9d3e03FC5BC175E47631Ff3F5d005ef268189",
+    "0x5C5585ab16ba26dCA3b698b8EA8eE6A5f01f0613",
+    "0x82A07DfF5B01d51B50176E8d9095A8cB8626b8bA",
+    "0x5bc91fCC71f04160E92d7a3430dadBb9365DBD4e",
+    "0xb6afA5C43bDb697051323C0e55d18365aBf8A64E",
+    "0xBa740B4C2F44639024e91cC0C9a6379a3fb89AeA",
+    "0x957705c7eeFB845b0a21A4085275D4fa36640dcd",
+    "0x19C5A59DA56D3bD4c4c3d9f64C594ACD2A9122D5",
+    "0x458c270f1AFa06035f477159F89707018d2Efb70",
+    "0xba7aA9a15C3f355b11EF9Fa0bA5e39dF4cdd6507",
+    "0x085211cf561C4196f8bc6dF368a4ca8EE7b2459a",
+    "0x9c61e1C5d636575a9D07d8A658fDC473F23BbbD0",
+    "0x282D55BA3932b11623E8Ab1Be51e572c4B2d0Dc7",
+    "0xF124a05AD0D63972FD76271d23C101106270910b",
+    "0x88335EC2258C89cB6A0d31b276360aDf14da6A54",
+    "0xF9C7F270b3A4EF08c630Cf447AA92969a8A3341a",
+    "0xbE4f6C928DDA7863B47e7eA81E61b1B8F8cC003a",
+    "0x179609e7865E4a3C4c668e9E851e9b689628dC5D",
+    "0xb2B20ca95616771Bb71F79c872274c5f626D3Fec",
+    "0x768FadD2F3e3BBfa721353E9bF84433D2095f014",
+    "0x768FadD2F3e3BBfa721353E9bF84433D2095f014",
+    "0x535eFd56D61F1f1E6f5A7c71CD578197d75cE780",
+    "0xb49a41656e0F6E34B256550251d35ae08028bbF2",
+    "0xe0982FbA9AF824585A7b4eA4D92C7BB816cC2669",
+    "0x5EC5AC30148e6Fad08408031C4e25DbD48646ee8",
+    "0x2ECA3BE891D210cFe8EF85505a3f66aeBe600775",
+    "0x8a1a76ff6e2a03f41317c77974FEa1d1246CB9cC",
+    "0x7f22788CCed30a13eb560714Fa8532976024bBfb",
+    "0xFa34245eF7721e7609D479A9991346A6121c25b0",
+    "0xd1B2a64a225e9BEf263f94654A47e28D514ccB87",
+    "0x6BAA71cE786F48022975280195F974634e90309e",
+    "0x6C38129583C485f5F18c4bAF59930CabDB60DCcC",
+    "0x3c9ac03AF6F34cc00b5e2a42b2010A69494C77D5",
+    "0x7447f2Bb45a6a3037e08FcCf303E4d3D1d68489F",
+    "0x65BB1615582BbeA7c383dB19624a5341D7dB8f8c",
+    "0x19f1f2837a4C8F1C0dE1342c54584480DC924a09",
+    "0x16497f0B98d9A0d14813C6ae0C948b1A0579Fb0c",
+    "0xc6F91ba8d86De4C9255B122045CDd7F89E540A8C",
+    "0x55ff391547f156B211a68e4787643989857C2C9f",
+    "0xB8D96AC691193c705D61dc4a2f7D15Daa259Ad0C",
+    "0x3Cef84E3AD485634d13397C2b6686963C4aDE5b9",
+    "0xF70fCa78fdF5fD29f72Dd665f46AEa1E6eb234C0",
+    "0xC2F4138B63694819B878d55b04eC0Cc33D3C0a82",
+    "0x8cc0A2197AfC7fC8607fC4802E693eAbB9A3a453",
+    "0x6839f8Ea0663e37E7145c6915f0F2E8fb211970B",
+    "0x2138Ea897550Ce8868Ee5fe2Ffc78407Ccd4387A",
+    "0x261b2A81D4a1d1cb17db9Cb11F5cacC49BC1c326",
+    "0x796A7051C83E7082309FfAFc56D03553443168ba",
+    "0xdc2e2ac75262204908c2e684D31914787883eF20",
+    "0x9a6ef672c9af8c98201D3DDfFBa9de4A67Bb7Df8",
+    "0xa93aD0E469B561d8B327D5d8CA0E6bF49479c6dE",
+    "0x3710d1ced6be5631195fdf80f433995f909e2d7f",
+    "0xf343847da9bac1215d400a51b303488155b08d0c",
+    "0xd8e40679ff3930265ff19aad301fd01a75e66006",
+    "0x55623b455238c9f7b56d97023aeff202dbeb4e96",
+    "0xd161f89f75e78b53edd546fcdfcc78a52575bfa8",
+    "0xc4a730dbebd1b3d93930427d6c2c5c25c7d01707",
+    "0xc2b8ab714fc050efc0d2ce5ca813eee01d3efd1e",
+    "0x3557c9d5263cfd9028089a796d4c865f0ed59505",
+    "0x71c5ddc4e61c501a34d388ca476919aca0cfecb0",
+    "0x6f9e8f0a4a85783db134641b538b8717def4a721",
+    "0x3b3199688d8ae08dce6cc1ea325db6ee1deecf23",
+    "0x0a6e5031054668f58d234372e5874007327c8dea",
+    "0x949710d6ff4d3d8030694b5582126bad57250066",
+    "0xEa009F5ce91e3F83399D98077147F7b1F4601db0",
+    "0xEa009F5ce91e3F83399D98077147F7b1F4601db0",
+    "0x7FC519Eb08323624767f8cE864B3810e9C9900f4",
+    "0x26eb596be49616B585eA9A6eE41a0A6F9975877a",
+    "0x06b7aD3cfC19eF98B8894CeAe985A6d1Adf6F424",
+    "0x3B91d6f5F08FC81Ef02443C6Be418cb7138816C9",
+    "0x2236F2C8d26c49A0f97C3B82F55cd60AFB72259e",
+    "0x151Aa631Bde3394501e72dd73f2F55F0b470f66b",
+    "0x2c7A715A27c63a91f9ECB4dB067eCF86170EBb30",
+    "0x3d04224b9Ad0Ad111a3bc63b96b0DF9D0C7fe212",
+    "0x1AD216eD17aaB9Fab83a5BC637fA384E81bfc708",
+    "0x78110D9338ed37F3D6582F9f98ad404a16f2AdAC",
+    "0xc1cEb68624d12E2810544f06b77cEf63C79083a1",
+    "0x67e6c5F218cB35c8078434e0f3DECEB6d582C6a9",
+    "0x1606976897bf57cd277Fa2968D9D168D1A49CC50",
+    "0xE8a72473c8f430eDCF645e57710031b025edFeCc",
+    "0x77EcD0017564D69bad99E2b09d0f0D93aea6C149",
+    "0xc7f9e50AF1e56523AfB6fBefcb2E9479FaC0D78C",
+    "0x888E10a00d2e880908b4445d40b058a80Ad658e8",
+    "0xc668c0789F0Ca00704B32B41b44694Be6E8154e1",
+    "0x2Ea6A9688A7Db690313ED087525741a67c5818A4",
+    "0xbf51086A27A7Ce74cE822F34a6Fa7ff34A4728a2",
+    "0xE2B725072ce13a0Fda3207450Fd9a50457B39BD8",
+    "0x3758189673c64635771F615e952EDFB062a24269",
+    "0x9d09E62588623Cf6E5535B62289E4A018d0D5E93",
+    "0x564Fe7Bb8752f61560f6410B4b73000a327aa89a",
+    "0x19D51E669073F7ee24C819bC2C1d55ce1df6aC6E",
+    "0x1f2a2cF6dB96431ce23fD8473583883983382d24",
+    "0xD80F3e3E2e4f7B03F7A112c65B1E8c0aa9754498",
+    "0x215d51844140535a0D93014eA0667f485bB86979",
+    "0x3eD89c818e710f23Dd54BbDab2EB7Fa5D5174E9b",
+    "0xb8c1ead0aed3979c6b25a477a2e2c5a4a35f2096",
+    "0x29e242b1357dfd29e05589d02fd51d51dd7fe5a8",
+    "0x5fcd3aaa8fb737f8aa13525ba3ed1bdb689afbe6",
+    "0x26454ea42ad2539df3c0bd04c11da8f660abc69c",
+    "0x65e8667765f0c152516005f68c9b68ffe27a710a",
+    "0xD61426B46278aEEDDaC699e98523Ea0063460C69",
+    "0x80B9A849a18AA403bFA8D2b28d016C02AeA65090",
+    "0xe28579cF2ED6040cC30340749a83A373158A435b",
+    "0xD4f05162681946D542e906555E42d231a2513e3C",
+    "0x66232ff08a39F8D0BF3464E08605b8D02469830e",
+    "0xe28579cF2ED6040cC30340749a83A373158A435b",
+    "0xD4f05162681946D542e906555E42d231a2513e3C",
+    "0x0dFbbfA069Ab9CA043f47903DEaCA7C5af28af38",
+    "0x2CB74268413585F1d6f02628909d8061d8db52Dc",
+    "0x92ace8AA3C573936EB8E9639489b7EB21fD1E001",
+    "0x16eF8286935ad3258F4Bfe262998DE8218B20478",
+    "0xb657Ff4f5916165D778cB87e894C36362709b474",
+    "0xCfAd591fF4B9f2C6e10376128eda7028C6E6535A",
+    "0x56B8E9E48D444138C345b5b5cb31fc243A88aaE8",
+    "0x8c2DDb7133863E6a339A93D749E4cd5673a991e2",
+    "0x129fd01d1acc7239e2bd5bed65e9e2ab2530cffb",
+    "0x640492a772af2f01a89f93a9930eb976ba6c11f5",
+    "0x6a0f4200bA32481a16d182454817E52588653D5c",
+    "0xd8E6DffB104b26a70A1eaE0f1440E5Ae70d3B77B",
+    "0x198695adA3c17e3c804c2Ea5233a7e4F3fA41484",
+    "0x727C61ec0d55B17230525E24F11Da706aCaEa69E",
+    "0x3Aef0F0E408B84f2261F0D946e73F60d15EAdFd4",
+    "0xc0E1a715AeCc5cAf3FeD5eDbe0D5F0F36159752D",
+    "0x107d159F88aa51259812047E470F0bDf7079dd2F",
+    "0x6131A461a421E40dFF262cBc587c2450f626eCAa",
+    "0x4f7268e6fc40BFD772CE02dA0d3ddf2d886b700B",
+    "0x49a6edF8ce35b005cA809eB06AF56Fff8BDfa076",
+    "0x1374bFb623511Dc4cd65CCc94E9495d19FaFfa56",
+    "0xd6D1e4296d0f61888F748b22D8b2a9747f83FA31",
+    "0x9Acb3912192f05AD4398381Ebb9b2fECb6B6657F",
+    "0xf57f8f44A33fAFAd26Ad6f0Ef7644Bf78deE1171",
+    "0x343f25462C41cd2Bc1C903Abb84212c49bCb9D0d",
+    "0x69d7C206034CcF70f19797F6164a1897037A0D7d",
+    "0xAeD497367e15E7548ca9Df60f0e5CC476f56E44a",
+    "0xe42EdA174519d23e0e0b49af9BA52bcEb976CbB6",
+    "0xcFfE53BF410f9867698964F61FF76dfA2C54deD9",
+    "0x846370d35f8047CB16b954204958895942869431",
+    "0x296070B3E76C1D45C902fd5BDf22bCEC77EBF653",
+    "0xa6D232829BdE024695595BC12A1Db4e112F1BD94",
+    "0x9C47eF6100738C96B5f1a343D6b768c430ECd55e",
+    "0x25Cd6944A88DEC8C278f412CcFBd48283B0a5638",
+    "0x501ccCD488Dbc8a9BaCFa1Ce75d8352046bfCBE9",
+    "0xDEe120C4687D0255D267f4154e068ba9ac9c4125",
+    "0xBa740B4C2F44639024e91cC0C9a6379a3fb89AeA",
+    "0x3FC6AfC823576d35F658C35732ac27ac0af51Dc7",
+    "0xfe5aFD4544D1c6378354adFA03A97e94A728795e",
+    "0xD0928503878DBFc91859a4f32b2bD2c004d34aa8",
+    "0x3b2cCE28546F46A9797C604D21E4fA4E57c9383C",
+    "0x0107802DF45667FAf4E1b4E4F8c857fA3Ac30dE5",
+    "0x65b2156eC07F0bfc087D41D256F6df1140423917",
+    "0x700cc1ad587e838740f3A306579bE3d60a8B4CAb",
+    "0x1384F35dc9ff3a08550778119c2AbC1abEb49E22",
+    "0x18719D53Bdeb6b32599C858235EAa1307a6b1E25",
+    "0x172719Cc32fdA085Df75564533502224B02737bd",
+    "0x1a55f23bdA72478296fa1cAC1DB10eDD304960D9",
+    "0x24576b3bFdE2FaB4D0b642eE5B26637c42b63729",
+    "0xA577B339fAA9DEdf4E1121AA3Fe1d7271b8184eA",
+    "0xd88347Cb0EC2e7735569d9ce4767D30d34703Def",
+    "0x65E99FF02B635E2F441C925FA551DdcbbB6d1fB5",
+    "0x809A53185EeFd79F44F54d310faBEB138432B40D",
+    "0x9Ce6DbE5fc6B4AcC00D2721Bc1CCa372D9bCC7e8",
+    "0x27B44238310dcb6b0A0b85515098eae3D32783A9",
+    "0x967023C6738C71d794731A96dC5C134bBbEC5622",
+    "0x7EEB0bbAC356646647954E7AaD89BA1b7eC271d5",
+    "0x1E312cB51F145bf9fCEA325e4AF039aA52ea75F8",
+    "0x0eC614CD67c71C97646510815F91488B5A934dA4",
+    "0x75B115D85AfDb0cCFa8Fa95aBb21c5f61D46efDC",
+    "0x810d4445D5a902b4BDca9ef048B1348d7eE90577",
+    "0x6da6400527eA88A8e8ef784427979D904c17F214",
+    "0xe606157DCf758998bbB5Be43Bd33F5133A8E653f",
+    "0xd85c11F1b48524b1e9B6Ff6569a37D172F7d60f2",
+    "0x694F42D3291d108a9135Ee4a2Bf547178E8E1452",
+    "0x816e69e29611C4A13D156e726F9C49D2877c1d75",
+    "0xAd41A16b30a622DEC378F21085A45e177F80ad90",
+    "0xfB341000Ad70CB6B49324Ee54792B66cb4e503C6",
+    "0x82D966baDe6f348D8afe01706890403384385C2B",
+    "0x2C5cF36d1488CF7Af68D52Ed6058e2BA3968d54e",
+    "0x1090e0789C412b71dd344aF00b15Dd04EAF25e73",
+    "0xbC2F05F49F9BD03Bd1b4212abC46f0b76215fDdC",
+    "0x609cBDdE43eCb23D83614343a4D6dD375eD8EE2F",
+    "0x1360cBDefeA06a729d4EdF078A62Ce8f13BfDc98",
+    "0xcB5F0F320d923880e09e9Dd93b029F2270E50bA0",
+    "0x76376460c1C4CedFb4DcCe3572D02DD7DF6844aC",
+    "0x8D9e8c79Fa12d289986361bDF7719c25420D0aBF",
+    "0x63489a9F83e627e5001C1882bBd2b9b9f3C82420",
+    "0xDFfCBbb056888C404640CE777c7e938966c38F0F",
+    "0x39F6273d75DB12f4E02957180AbEaE9132800d17",
+    "0x90C4fCDc6Ab5Bc8aAc13eA1d694D2A19f175625f",
+    "0x9701824C0D3bd2146805c3A51FC5E747b44b0fCE",
+    "0x7baD7aABa16736759FE53ce8e38eAD8837da502f",
+    "0x38a55688B5b6a60690EB83aE4dDD6b763C0dA487",
+    "0xcbBdADa0E2F60400e64eC81647Ed730C16227e71",
+    "0xc71dBA523b29952487aA2F8d95c3c5cFF29C26B9",
+    "0x3acA574627afd033BBA1839d12b2Fa0f47e04c55",
+    "0x6F3D1c85c405fc4D37D3D31e6C8a981dE93E950e",
+    "0xF5c475B83B87973be173Df3DC4eD8383C42A39A9",
+    "0xde7FDdD94170995B9524eF178bfD17eb149991E8",
+    "0xd76c74f205976cf2e504baa3d7d8d828c1164132",
+    "0x29af1814594eb9b23d4aef6fb640fcc5b46088e6",
+    "0xDCAc0ee047DA0aB449f6bEb08c8f79c00eaBC67D",
+    "0xf6b0823F7D07301e58b1E808B18F29442a735E63",
+    "0x013Ec3eD2A613644D71F0d7d9376a6AE4b3d63c2",
+    "0x80cF051A318762a8cC97Ff95514d72436a226511",
+    "0x8Bd3E54eC95fBFB397E37c73390c6D68f2dF4718",
+    "0xFC29736aC7B8Ac8106D24515312396093d3D2141",
+    "0xD26E54e794e6A1102218FeCEdCd7BF26336228f0",
+    "0x0Ede308585E453dfa428fe974741f06f76610e8D",
+    "0x9F7a1dA3bc621A0aaeB83eeF3810c972d8C3B68e",
+    "0x38E11A6E34A92EcFF736bC9E142DFA15025De301",
+    "0x7e596acdbe4552a4e64b9979a90be788dff63be5e15",
+    "0x3bA8fdfda5b5afd759804a8494f88c00CC283b81",
+    "0x5413b45F30C3a1373d9594Fd8BBE22aC6cd7dbBf",
+    "0xEF28629B04fa176C3366619dc6eBbffE4fbb5196",
+    "0x7Fb633b9d378D6ABbc26E900b61aA64E014486E1",
+    "0xA5989632ccD56ea38F2BEe349448E1b4Fc5ac7cb",
+    "0x4d594054b58F20A4ae89A176ADC93d0079cF7E8f",
+    "0x43a9421BE4adFCA3D3cdab8cb0ea80015af2CAf9",
+    "0x20942503ceD311884157A1EEdD3351350D246978",
+    "0x1a6C65aeD06c0E4FAaB2384F2b13Fc520365B9f0",
+    "0xca8f94B58072a3C9A8dC7131Ff389Fc66A0d1609",
+    "0x0C2ABA1Eec154bfcaf41bB5bca756caf5BEcea70",
+    "0x1C1BDce70EF4CdaD6A0C889Ee69AEFca366D1365",
+    "0xA40A549d7B9D702F24C7Ed47e5766C6e60754067",
+    "0x2468639Ae120860661F05E697784988D2D2677f4",
+    "0xDe82fA3fC9b6A7b99803D99F3C0bdB93A4ba045f",
+    "0x99B0740d0f68F31E5d46f48486a06484778D12F3",
+    "0x9a8D796723fA9D288F9670aF281b8CE57eEE1A75",
+    "0xfC62685C0F536853daD69D98988A59ee46AA89be",
+    "0x66Bb6cb049DCf436a8a4D1f2cEb2A162e90f82cf",
+    "0x853bD0D5b00BaFAb1E30Fa0d540a07f3D343C56d",
+    "0xd731faF4aaa69950d8fDb9A5ADAD0Df32Ec39773",
+    "0x923546659A027979AEeDeE7a25FF6c990ea0B03a",
+    "0x53fBd6c5aCb16136E7392081233C425E3A533086",
+    "0x53fBd6c5aCb16136E7392081233C425E3A533086",
+    "0x0EBF6a127aB0350342303b6670AF1E70bAC280b4",
+    "0x7f411bf94c830a3ed1fa3691025a92a4a863f824",
+    "0x5ae023ede4fe0c0f29ca9efa7dba94f23dc14795",
+    "0x27fcb3903aca8416f4308709fe6e0d0123a55661",
+    "0xaeb771bcee332e01baa775aba910613b05ed697c",
+    "0xfe65b5457a4baf1a76311dbefa5397c51b2b2782",
+    "0xA0B6eE3958E3115073d4cB8191D4a9D45505225D",
+    "0x6c5760AFA61aC14275B42d9798Ae44c5e89af1c6",
+    "0x83c9bB63110b974C7c85EDaeFE229EbA01af88ce",
+    "0xC2d66DF63ac6B2822Fa038E85D0431182bDF3F96",
+    "0xcd759427C20745E8E38259CB41550dC76C89D2c9",
+    "0x37Cf173047f9f69CfD6a4f1dFbF012ac3c99452B",
+    "0xF855136c1969c45E0F2F9fEf0a9Fd8b995b15deE",
+    "0xD8880B5bC216E7dB153e34aDc21b36F69B7926A4",
+    "0x866f5e76B4A60982028610B9c0289E54460DD6b9",
+    "0x5FCB318D23e17EB5Fc3fdC7388EDe235bf34523A",
+    "0xd6b2cCe5C0480780F3e130925FDc823B65aA6C5B",
+    "0xc2BC66fab4398Bae35b941CbB14eB3B6584fdCcb",
+    "0xf6c211D6270987FBcB869Bd93c56450b49167782",
+    "0x66298f9dFA4b65F77494220f6c5AF18CE78E8a50",
+    "0x8E3a9013b3e5C522fDd841d1ffe685129c876187",
+    "0x9984E0CB3Eda7251CC17cD1E19Ff1CCe0121207C",
+    "0xdA5D958dDF921F07a3C465978B340228692e5761",
+    "0x483548f972326A8729afF0761864243f9b0a6Be4",
+    "0x3bd9343a035162e05CF363A53F7e3aAcb7A85C91",
+    "0x0A9cd0de60cB0BC3DB3730Cb3b352C722EEafd7D",
+    "0x75b8FBE6b8a88C4E9c747A05CE9dca570bD7f626",
+    "0xEa27Bb66677851478433E73a1A0195F9b2251265",
+    "0xe300c55d990e50DD740A04A33fF9Bd68CE9BfE08",
+    "0x9C0f983665C863d528C22ca1fCAED86E9DF22E07",
+    "0xE1B4d5009273B1c4A27cb1630cf8E46C5A0783d3",
+    "0x88186216f4F54BcEC29fea24A1e4dF1e6D3492B1",
+    "0xEd90a5a9Bc225205946877A5650C83E00013F640",
+    "0x5Ed0691E14EDeD0164407F93E160cd0ee68a9b57",
+    "0x2BD6194613De4FC6B5Cc952f639EAfCCAe1794f6",
+    "0xD1c887B4b2ecac8E3FCACa62B2cB6A3d9aC75Aeb",
+    "0xC186f8e7C4515C2fAC278f60734D0a4fA788BA97",
+    "0x68704beCFBb41d8E7eB110f18118E94cDBC841B8",
+    "0x6e6435607AaF2146687CCFB037ee841E95747791",
+    "0xeB2cF592De5AEa5f72e406c6255C75a98f951934",
+    "0x7d6F656B8F19e9cD81420aa63A0Ad79d9Ec5B468",
+    "0x49EB2165a220617Ff044Aa72F0177407e79677D1",
+    "0xeF0928128c63C7B2d47e83cb896Ac7C99B51Fb0F",
+    "0x78D53975bf67eC79D675599b4a90398F5B214b4A",
+    "0xA190703A5ADb63CcBED334f5a4e365E8dc0d5d43",
+    "0x3dF3D254c4c16fca37a50A82eA5a28F1192D2a16",
+    "0x49EB2165a220617Ff044Aa72F0177407e79677D1",
+    "0x018E497E8a116F914A76a8Ad0b2937A86A342a31",
+    "0xe1e4e0D2D5965c6dCC1c33C1637f2abf04535877",
+    "0x5d2981c7D09B0B24177D55866cB58D5904C7Fc31",
+    "0x240627127eF77d9cB2d5F9C472DD491a1C70B642",
+    "0x259FEb458e2A969c981e58BAdC2406C9303E10B0",
+    "0x95eA73cE923fE95b500c1AD3e7d7C9828226A097",
+    "0x911D983dC8f76b5B0C5d4beda8f4E907012AD463",
+    "0x6dE3f181B38acdCB85038D45Fd97F4299c9c10C2",
+    "0xa66c38396B9b5FB1e7ba9a240bb7224FfE2945B0",
+    "0x22fBCBB056fB83cefE4f3d66c14A8aF52149223F",
+    "0x166Be543704529ccA1cb9Bf06fdF0DdB04D83d63",
+    "0x01147EF4cF7f786FCB7711A40a3D046289794733",
+    "0x18cd142dD5c4dDfB756bcCdBf5C4eaE5dFF403A5",
+    "0x8F6f8Bdb7fcfD1AbEa0d079b24Cc3789f0D66446",
+    "0x7Be28414FB12906f23f73E65f85aF526eBafF949",
+    "0x4f33056CBc5Ac057A0C6ecc4272717e4f10920E9",
+    "0x982C35DD9291c83bF1CdEe443fe368d4BBd9d1dA",
+    "0x85F1f4FEF2aaE59D7E15716d7a89Da8E48B6E817",
+    "0x86fcb5c610c0ccf6c18E0Bd3F56ee696689ECe02",
+    "0xF9316b784Aa52C90B5618EE284c93D546C6F4737",
+    "0x3A3f27A11BF305f48c2A057A8F0f21b76C24bf06",
+    "0x85D52f32338E6E0ad60Ac5DdA367098bf88D7a21",
+    "0xa685a756431D370c5168f843891E5D4E0dC21b33",
+    "0xe36E95e7d4981Af12586Ea28bF794C85d3511619",
+    "0x236AeB0245169a44F623b01654EC749dcbFAe6c7",
+    "0x063EA9A84b21e8862E2EB0193FAbAcc9c7ae5A49",
+    "0x3Fd2782efc4F9cac4146Ca5fA11354f26371444b",
+    "0x4957E098587a5e848885c3d10A1eFf451c77eD34",
+    "0xAD4e1879297249d43Abd7A831D03d6D7Eb471B2F",
+    "0x71e9Ddf441a60f789E309fFCf7373303D16830Fb",
+    "0x2a0838672A856ec6b279c0c68917393D6C9FFf6d",
+    "0xfbb5B096AB4605f7dCc2e139033919Bd9856fADF",
+    "0x809bCd995860b038b32a8cE30685F31A8C2911C7",
+    "0x9e3E9e37d763D7e06c9346ecA520FCCf40672Ef9",
+    "0x2C0E1A630Dd7E3510d0b76BA9Db98a05f309A782",
+    "0xC4439c29F6b83e5C3449C071fC305B34eF0BeAe3",
+    "0x82B1fa9c9909c3ac0A616dfF2F29eA360b287760",
+    "0x3Ade4f43FAafa1B4491B03c2d3eD343fEb3917c5",
+    "0xA7C86acD2D1B22AA7a106D5B4B603Fceb3Ff1FA1",
+    "0x792C65601c70Be476e03DacE173Cdb1Bc72c1F76",
+    "0x085aD778a60d405014aA98B8908518bc861D7165",
+    "0x1071bBfAa600A96DC1F396D0d941A91427575991",
+    "0x8C6eeFBD087cDD7f8906Ce74A26c968bC8Bd7e46",
+    "0xFC3733587E4066D664963c90CA3bAF76ff05fE46",
+    "0x0Da5eE7fBda8C018ba288cceAb4e8C83d309b27a",
+    "0xA4eE2105030eB2FF8239E8D3b97843f76487fbfF",
+    "0xF7f6C7a545059F06569aa95acC045C9752B56241",
+    "0x92325f514E3a4A1bD871752eC041793Adb885e52",
+    "0x93eA8a3612c7657976349B1e8D6a50905aFC1847",
+    "0x93eA8a3612c7657976349B1e8D6a50905aFC1847",
+    "0xa6183330bFB2b0C08a161C57B91C6127Af6D6803",
+    "0xC5261431ee26f77Ca70441cdDCaBC32F6792416d",
+    "0x98F00f84a49309E1b9644510D2a8A7c8317DBD94",
+    "0xD6992429A118920490951B43FF45Dbf0168C875C",
+    "0x3AACd6B0798F1Cb2f2183280d81470705Cb7ea87",
+    "0xeC9c698f15C94b812e414DD7b4C87543aEBe73aF",
+    "0xB435294442398fBb52780Cc8B115f09F85d27892",
+    "0xb7D20456050bb3e4c9416CA9a51e07aE0374876D",
+    "0x810101076b702A07984E73FD35fd6A04108F4d36",
+    "0xab4be46d9c49C3869cDbc0392acA7f5f93F3c09a",
+    "0x235275F125f55Ea3a209e6B1847de5a606a3D877",
+  ];
+
+  console.log(balanceOfCheck, "the balance");
+  console.log(balanceOfCheck2, "the balance");
+  console.log(address, "the address");
+
+  //converting hex to normal number // note - .toLocalString("en")
+  const hexToDecimal = (hex) => parseInt(hex, 16);
+  // const bal = balanceOfCheck
+  // console.log(bal)
+
+  const balance = balanceOfCheck ? hexToDecimal(balanceOfCheck._hex) / 1e18 : 0;
+  const balance2 = parseInt(balanceOfCheck2, 16);
+  console.log(balance);
+  console.log(balance2);
+
+  // if user wallet is not connected, this function will be called to the user attention
+  const connectWalletError = () => {
+    if (!isConnected) {
+      toast.warning("Please connect your wallet first to continue", {
+        position: toast.POSITION.TOP_LEFT,
+        theme: "dark",
+        autoClose: 5000,
+      });
+    }
+  };
+
+  // if user wallet is not in the WL, the warning appears. If they are they proceed to dapp
+  const enterDapp = () => {
+    const lowerCaseAddress = address.toLowerCase();
+    const lowerCaseWhitelist = wl.map((wallet) => wallet.toLowerCase());
+
+    if (lowerCaseWhitelist.includes(lowerCaseAddress)) {
+      navigate("/airdrop-bot");
+    } else {
+      toast.warning(
+        "Your wallet is not whitelisted. Please proceed by whitelisting your wallet address.",
+        {
+          position: toast.POSITION.TOP_CENTER,
+          theme: "dark",
+          autoClose: 5000,
+        }
+      );
+    }
+  };
+
+  return (
+    <div className="page-container">
+      <Helmet>
+        <title>alfa.dapp</title>
+        <meta property="og:title" content="alfa.dapp" />
+      </Helmet>
+      <ConnectionButton></ConnectionButton>
+      <div className="page-banner">
+        <img
+          src="/playground_assets/dappper.svg"
+          alt="image"
+          className="page-image"
+        />
+        <h1 className="page-text">alfa.dapp</h1>
+        <span className="page-text1">v1.0</span>
+        <span className="page-text2">
+          You need to whitelist your wallet first.
+        </span>
+        <a
+          href="https://whitelist.alfasociety.io/"
+          target="_blank"
+          rel="noreferrer noopener"
+          className="page-link"
+        >
+          Whitelist Wallet Address
+        </a>
+        <button
+          className="page-button1 themebutton button hover-white"
+          onClick={!isConnected ? connectWalletError : enterDapp}
+        >
+          enter dapp
+        </button>
+      </div>
+    </div>
+  );
+};
+
+export default DappGate;
